doc and debug connectors
